LinkedinRE/requirements.txt
langchain
openai
tiktoken
feedparser
bs4
firebase_admin
python-dotenv